Split public/private API
